Use LR4J API for junit tests
Changed to use the LR4J API to output recordings on junit test failures
instead. This produces much smaller recordings and runs the tests more
quickly too. Note that the "-Dundo.basedir" and "-DargLine" properties
must be supplied as mentioned in the readme because the LR4J API
compile and runtime need to know the location of the libs.